main/postmarketos-ui-xfce4: clean up post-install script
Move the line about the LightDM autologin user out of the -openrc
subpackage post-install script into the package post-install script,
as this is independent from OpenRC.

While at it, in the post-install script also remove the line on tinydm
MR 3887. This is more than a year old and part of the last stable release.
Therefore it can be removed by now.