main/postmarketos-ui-lxqt: save and restore brightness on boots
LXQt can control the display brightness. It doesn't save and restore it
over boots, however. In distributions with systemd this is handled by
"systemd-backlight@.service". For OpenRC the package brightnessctl offers
a service saving the state of the brightness at shutdown and reboot.

The brightness save/restore function of brightnessctl can be disabled
by the user by disabling the according OpenRC service. To avoid it being
re-enabled on every postmarketos-ui-lxqt package upgrade, the enabling
of this service should not be part of the post-upgrade script permanently.
Adding it temporary could be an option but maybe it's better to keep it out
of the post-upgrade at all to avoid a too large unexpected change in behavior.
The service will therefore be enabled on new installs only.

Also the other contents of the current post-upgrade script (a symlink to the
post-install script) are not neccessary on upgrade, therefore remove the post-
upgrade script entirely.