systemd/systemd-services: manually issue chvt on start of phosh (MR 5757)
libseat 0.9.0 no longer performs an automatic VT switch when taking
control of a seat to allow sessions to be started in the background. As
consequence, one must manually switch if one wants a background session
to be immediately activated.

Issue a chvt on ExecStartPost, when the service is deemed running.

Upstream: https://gitlab.gnome.org/World/Phosh/phosh/-/merge_requests/1566
[ci:skip-build]: already built successfully in CI