temp/u-boot-qcom: new aport
This forks the u-boot package from alpine, targeting the qcom-next
branch and provides a generic qcom binary.

For debug/serial output are soc-specific subpackages with uart
configuration is provided.

For each supported device is additionally a boot.img generated
(both the general qcom binary and the soc-specific debug ones).
This works together with "fastboot-bootpart" for now.
My hope is that we find a solution where fwupd is taking care of
existing installs of u-boot and some helper (perhaps within
pmbootstrap) helps installing this second-stage bootloader.