WIP: linux-next: keep old versions on upgrade
Add pre and post-upgrade scripts which keep the old kernel, modules,
dtbs, and release file.

This allows for the kernel package to be upgraded while preserving the
old version, so if the upgrade causes a regression the user can switch
back.

This will need to be heavily refactored, documented and made generic.

Most likely the upgrade scripts will be moved to some -dev package and
the kernel package APKBUILD will simply source a script to create the
pre/post-upgrade scripts.

For this to work, the kernel release directory
(/usr/share/kernel/something/) has to follow the format
<flavour>-<pkgver>-r<pkgrel> since the scripts rely on this to map the
package version to the kernel suffix.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>