Elementary pantheon WIP w/ systemd
Doesn't get very far yet... The pantheon greeter is totally broken.

lightdm PAM config needs to be adjusted to reference pam_systemd.so
instead of pam_elogind.so (maybe we should add a symlink?).

If you make it to the desktop by using the GTK greeter (and choose
"secure session" for wayland) then you'll just get a wallpaper and a lot
of kernel messages about wingpanel crashing with SIGTRAP.

Haven't debugged further.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>