device-motorola-smi: remove
This testing device is the only user of petitboot, which fails to
build. Remove it from the stable branch for now. Somebody can bring it
back later if they like - make MRs to fix it against edge first and ask
for a backport to v24.12. I'll make similar commits following this to
drop other packages that don't build.