systemd/abuild: drop
we have abuild in both the systemd and edge repos due to #3452

it seems like just adding an explicit bootstrap step is enough to get
things going properly. That probably isn't even needed since it seems
like the bootstrap failure we saw was specific to that scenario and
probably wouldn't show up again. but better safe than sorry.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>