**/{linux,u-boot}-*: use pmb:cross-kernel instead of cross-native
The way we cross-compile kernels is actually incompatible with abuilds
own cross compilation support.

To support actually cross-compiling other packages (by integrating mesons
own cross compilation support for example) it's necessary to correctly
configure abuild which breaks building the kernel and u-boot.

Tackle this by introducing a new pmb:cross-kernel option which preserves
the old cross compiler behaviour so that we can upgrade the
pmb:cross-native option so that it enables proper cross compilation with
abuild.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>