main/postmarketos-config-nftables: add rule for ausweisapp2
In the process, bump the pkgver to 1 similarly to what we do with
device packages where version numbers don't have a meaning. We
specifically don't create a subpackage, since the app is in flathub,
which is officially supported by the provider, and is a common way for
most non-technical users to ust it, since it's offically endorsed.

Unconditionally opening this really high port should also not increase
the attack area in any remarkable way.