CI: Ensure upstream remote is not shallow
Otherwise this can lead to `merge-base` not finding a merge base, and
then stuff *goes wrong* with CI.

So when is this relevant? In the situations where the gitlab runner's
git repo is cached between runs. This is a real pain and I don't get how
they thought it would be a good idea.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>