main/postmarketos-mkinitfs-hook-ci: Print markers many times
This helps ensure the logs are actually flushed.

The numbers chosen are relatively arbitrary, and it might be that using
a longer delay, maybe with more messages too, is desirable.

Right now, it's printing 20 times over a span of 2 seconds.

The whole `precision` schtick is how we can approximate float division,
when using only integer division. Then, the important trick is to shove
the given number back into the correct representation for `sleep`, and
careful `printf` application gives us exactly what we need.

Signed-off-by: Samuel Dionne-Riel <samuel@dionne-riel.com>